Finish map function and fix bug

+++ app/src/main/java/edu/hcmuaf/tourrecommendationapp/model/Location.java
@@ -25,7 +25,7 @@ public class Location implements Serializable {
      */
     private double locationLongitude;
 
-    private int distance;
+    private long distance;
 
     private int wishListOrder;
